build: Switch scripts to .mjs.
