--- bin/generatePo.js
@@ -1,6 +1,0 @@
-#! /usr/bin/env node
-/* eslint-disable */
-
-const { generatePoFiles } = require('../packages/generator');
-
-generatePoFiles();
