+++ bin/generatePo.mjs
@@ -1,0 +1,5 @@
+#! /usr/bin/env node
+
+import { generatePoFiles } from 'emojibase-generator';
+
+generatePoFiles();
