--- bin/generateShortcodes.js
@@ -1,6 +1,0 @@
-#! /usr/bin/env node
-/* eslint-disable */
-
-const { generateShortcodes } = require('../packages/generator');
-
-generateShortcodes();
