+++ bin/generateShortcodes.mjs
@@ -1,0 +1,5 @@
+#! /usr/bin/env node
+
+import { generateShortcodes } from 'emojibase-generator';
+
+generateShortcodes();
