--- bin/generateTypes.js
@@ -1,6 +1,0 @@
-#! /usr/bin/env node
-/* eslint-disable */
-
-const { generateTypes } = require('../packages/generator');
-
-generateTypes();
