+++ bin/generateTypes.mjs
@@ -1,0 +1,5 @@
+#! /usr/bin/env node
+
+import { generateTypes } from 'emojibase-generator';
+
+generateTypes();
