+++ package.json
@@ -12,12 +12,12 @@
     "coverage": "yarn run jest --coverage",
     "docs": "cd website && yarn run start",
     "lint": "beemo eslint",
-    "generate": "yarn run generate:emojis && yarn run generate:shortcodes && node ./bin/buildFilesizeTable.js && yarn run generate:tests",
-    "generate:emojis": "rm -rf ./cache/final && node ./bin/generateEmoji.js",
-    "generate:po": "node ./bin/generatePo.js",
-    "generate:shortcodes": "node ./bin/generateShortcodes.js",
+    "generate": "yarn run generate:emojis && yarn run generate:shortcodes && node ./bin/buildFilesizeTable.mjs && yarn run generate:tests",
+    "generate:emojis": "rm -rf ./cache/final && node ./bin/generateEmoji.mjs",
+    "generate:po": "node ./bin/generatePo.mjs",
+    "generate:shortcodes": "node ./bin/generateShortcodes.mjs",
     "generate:tests": "cp ./packages/data/en/data.json ./packages/test-utils/test-data.json && cp ./packages/data/en/shortcodes/emojibase.json ./packages/test-utils/test-shortcodes.json && cp ./packages/data/en/meta.json ./packages/test-utils/test-meta.json",
-    "generate:types": "node ./bin/generateTypes.js",
+    "generate:types": "node ./bin/generateTypes.mjs",
     "test": "beemo jest",
     "format": "beemo prettier",
     "prerelease": "yarn run pack && yarn test",
