+++ packages/generator/src/helpers/readCache.ts
@@ -2,7 +2,7 @@ import fs from 'fs';
 import path from 'path';
 import { log } from './log';
 
-export function readCache<T>(name: string): T | null {
+export function readCache<T>(name: string): T | undefined {
 	const cachePath = path.resolve(process.cwd(), 'cache', name);
 
 	if (fs.existsSync(cachePath)) {
@@ -13,5 +13,5 @@ export function readCache<T>(name: string): T | null {
 
 	log.info('cache', `No cached data found for ${name}`);
 
-	return null;
+	return undefined;
 }
