+++ packages/generator/src/loaders/loadShortcodes.ts
@@ -1,7 +1,7 @@
-import path from 'path';
 import { ShortcodeMap } from '../types';
 
 export async function loadShortcodes(): Promise<ShortcodeMap> {
-	// Relative to the lib/ folder!
-	return Promise.resolve(require(path.resolve(__dirname, '../shortcodes.js')) as ShortcodeMap);
+	const { shortcodes } = await import('../resources/shortcodes');
+
+	return shortcodes;
 }
