+++ packages/generator/src/resources/shortcodesLegacy.ts
@@ -35,8 +35,7 @@
  *    https://www.dailywritingtips.com/100-words-for-facial-expressions/
  */
 
-// eslint-disable-next-line import/no-default-export
-export default {
+export const shortcodes = {
   // 🇦 regional indicator A
   '1F1E6': ['regional_indicator_a'],
   // 🇧 regional indicator B
