+++ packages/regex/emoticon.js
@@ -1,2 +1,1 @@
-module.exports =
-	/[:=](?:\'(?:\-[\(\)DO\[\]o\{\}]|[\(\)DO\[\]o\{\}])|\{>|\-[#\$&\(-\*\/3<>-@B-EJLOPSXZ-\]cjlopsxz-\}]|[#\$&\(-\*\/3<>-@B-EJLOPSXZ-\]cjlopsxz\|\}])|[Oo][:=](?:\-[\)\]\}]|[\)\]\}])|>(?:[:=](?:\-[\(\)\/\[-\]\{\}]|[\(\)\/\[-\]\{\}])|0(?:\-[\)\]\}]|[\)\]\}]))|%(?:\-[\(\[\{]|[\(\[\{])|\\[Mm]\/|D(?:\-[:=Xx]|[:=Xx])|8(?:\-[#\)D\]\}]|[#\)D\]\}])|;(?:\-[\)P\]p\}]|[\)P\]p\}])|x(?:\-[\(D\[op\{]|[\(D\[op\{])|X(?:\-[\(DOP\[\{]|[\(DOP\[\{])|<\/?3|[:=]\{/;
+module.exports = /[:=](?:\'(?:\-[\(\)DO\[\]o\{\}]|[\(\)DO\[\]o\{\}])|\{>|\-[#\$&\(-\*\/3<>-@B-EJLOPSXZ-\]cjlopsxz-\}]|[#\$&\(-\*\/3<>-@B-EJLOPSXZ-\]cjlopsxz\|\}])|[Oo][:=](?:\-[\)\]\}]|[\)\]\}])|>(?:[:=](?:\-[\(\)\/\[-\]\{\}]|[\(\)\/\[-\]\{\}])|0(?:\-[\)\]\}]|[\)\]\}]))|%(?:\-[\(\[\{]|[\(\[\{])|\\[Mm]\/|D(?:\-[:=Xx]|[:=Xx])|8(?:\-[#\)D\]\}]|[#\)D\]\}])|;(?:\-[\)P\]p\}]|[\)P\]p\}])|x(?:\-[\(D\[op\{]|[\(D\[op\{])|X(?:\-[\(DOP\[\{]|[\(DOP\[\{])|<\/?3|[:=]\{/;
